test(contract): add deploy and call contract function celo test
