+++ README.md
@@ -1,6 +1,6 @@
 # <h1 align="center"> ethers.rs </h1>
 
-**Complete Ethereum wallet implementation and utilities in Rust**
+**Complete Ethereum and Celo wallet implementation and utilities in Rust**
 
 [![CircleCI](https://circleci.com/gh/circleci/circleci-docs.svg?style=svg)](https://circleci.com/gh/circleci/circleci-docs)
 
@@ -20,6 +20,24 @@ ethers = { git = "github.com/gakonst/ethers-rs" }
 
 </details>
 
+### Celo Support
+
+[Celo](http://celo.org/) support is turned on via the feature-flag `celo`:
+
+```toml
+[dependencies]
+
+ethers = { git = "github.com/gakonst/ethers-rs", features = ["celo"] }
+```
+
+Celo's transactions differ from Ethereum transactions by including 3 new fields:
+- `fee_currency`: The currency fees are paid in (None for CELO, otherwise it's an Address)
+- `gateway_fee_recipient`: The address of the fee recipient (None for no gateway fee paid)
+- `gateway_fee`: Gateway fee amount (None for no gateway fee paid)
+
+The feature flag enables these additional fields in the transaction request builders and
+in the transactions which are fetched over JSON-RPC.
+
 ## Features
 
 - [x] Ethereum JSON-RPC Client
@@ -28,6 +46,7 @@ ethers = { git = "github.com/gakonst/ethers-rs" }
 - [x] Querying past events
 - [x] Event monitoring as `Stream`s
 - [x] ENS as a first class citizen
+- [x] Celo support
 - [ ] Websockets / `eth_subscribe`
 - [ ] Hardware Wallet Support
 - [ ] WASM Bindings
