+++ ethers-contract/Cargo.toml
@@ -22,7 +22,8 @@ futures = "0.3.5"
 ethers = { version = "0.1.0", path = "../ethers" }
 tokio = { version = "0.2.21", default-features = false, features = ["macros"] }
 serde_json = "1.0.55"
+serial_test = "0.4.0"
 
 [features]
 abigen = ["ethers-contract-abigen", "ethers-contract-derive"]
-celo = ["ethers-core/celo", "ethers-providers/celo", "ethers-signers/celo", "ethers/celo"]
+celo = ["ethers-core/celo", "ethers-core/celo", "ethers-providers/celo", "ethers-signers/celo"]
