+++ ethers/Cargo.toml
@@ -32,7 +32,7 @@ celo = [
     "ethers-core/celo",
     "ethers-providers/celo",
     "ethers-signers/celo",
-    "ethers-contract/celo"
+    "ethers-contract/celo",
 ]
 
 core = ["ethers-core"]
